style(Home.js): make homepage responsive and refactored code
