+++ components/home_sections/ArtForYou/ArtForYouHeader.js
@@ -1,0 +1,7 @@
+import SectionHeader from "../SectionTemplate/SectionHeader";
+
+export default function ArtForYouHeader(){
+    return(
+        <SectionHeader headerText={'ART FOR YOU!'}/>
+    )
+}
