+++ components/home_sections/ArtForYou/ArtForYouSection.js
@@ -1,0 +1,11 @@
+import SectionTemplate from "../SectionTemplate/Section";
+import ArtForYou from "./ArtForYou";
+
+export default function ArtForYouSection(){
+    return(
+        <SectionTemplate
+            renderSection={<ArtForYou/>}
+            height={51}
+        />
+    )
+}
