+++ components/home_sections/FeaturedArtists/FeaturedArtistsHeader.js
@@ -1,0 +1,7 @@
+import SectionHeader from "../SectionTemplate/SectionHeader";
+
+export default function FeaturedArtistsHeader(){
+    return(
+        <SectionHeader headerText={'DISCOVER ARTISTS'}/>
+    )
+}
